merefactor code dan menambahkan dokumentasi API

+++ BE/controllers/chapterController.js
@@ -21,7 +21,7 @@ const createChapter = async (req, res, next) => {
     });
 
     res.status(201).json({
-      status: "success",
+      status: "Success",
       message: "Sukses membuat chapter",
       data: chapter,
     });
@@ -47,7 +47,7 @@ const getAllChapter = async (req, res, next) => {
     }
 
     res.status(200).json({
-      status: "success",
+      status: "Success",
       message: "Berhasil mendapatkan data chapter",
       data: chapters,
     });
@@ -67,12 +67,12 @@ const getChapterById = async (req, res, next) => {
       ],
     });
 
-    if (!chapter || chapter.length === 0) {
+    if (!chapter) {
       return next(new ApiError("Chapter tidak ditemukan", 404));
     }
 
     res.status(200).json({
-      status: "success",
+      status: "Success",
       message: `Berhasil mendapatkan data Chapter id: ${id};`,
       data: chapter,
     });
@@ -121,12 +121,13 @@ const updateChapter = async (req, res, next) => {
     }
 
     res.status(200).json({
-      status: "success",
-      message: `Berhasil memperbarui data chapter id: ${id}`,
+      status: "Success",
+      message: `Berhasil mengupdate data chapter id: ${id}`,
+
       data: updatedChapter,
     });
   } catch (error) {
-    next(new ApiError(error, 500));
+    return next(new ApiError(error, 500));
   }
 };
 const deleteChapter = async (req, res, next) => {
@@ -136,6 +137,12 @@ const deleteChapter = async (req, res, next) => {
 
     if (!chapter) {
       return next(new ApiError("Chapter tidak ditemukan", 404));
+    }
+
+    if (chapter.modules && chapter.modules.length > 0) {
+      return next(
+        new ApiError("Tidak dapat menghapus chapter dengan modul terkait", 400)
+      );
     }
 
     const isChapterDeleted = await Chapter.destroy({
@@ -148,7 +155,7 @@ const deleteChapter = async (req, res, next) => {
     }
 
     res.status(200).json({
-      status: "success",
+      status: "Success",
       message: `Berhasil menghapus data Chapter id: ${id};`,
       data: Chapter,
     });
